… datafusion/core/tests/sql/order.rs
… datafusion/core/tests/sqllogictests/test_files/aal.slt
… datafusion/core/tests/sqllogictests/test_files/window.slt
… datafusion/optimizer/src/simplify_expressions/expr_simplifier.rs
… datafusion/sql/src/statement.rs
… datafusion/sqllogictest/test_files/aal.slt
… datafusion/substrait/src/logical_plan/consumer.rs
… datafusion/substrait/src/logical_plan/producer.rs