## Which issue does this PR close?

Related to https://github.com/apache/arrow-datafusion/issues/7196

This PR contains a proof of concept for a general purpose TopK operator -- I do not intend to merge this PR as is, but rather use it to gather feedback / refine the idea. 

When we are happy with that, I will prepare an actual PR for consideration and review

Specific questions for reviewers:
1. Do you know of a better structure than is in this PR (sorted `Vec`) for keeping a topk heap?
2. Do you have thoughts / opinions on a `TopKExec` (vs reusing `SortExec`)?

## Rationale for this change

The idea is that when sorting with a limit DataFusion currently spends 

1. a lot of memory retaining lots of data which will eventually be discarded
3. more CPU time than necessary to sort / merge rows which will ultimately be discarded

## What changes are included in this PR?

This PR contains a single TopK implementation, that uses the RowFormat. This new TopK is used in place of `SortExec` with a `fetch` in this PR, though I think it should eventually get its own `TopKExec`

Note I thought about using a specialized implementation for primitive types but didn't do it because:
1. The overall time is now pretty small (in my mini benchmarks the row format takes in total 2-3% of the time)
2. It would take some non trivial complexity to handle `nulls first` / `nulls last` and `asc` / `desc` correctly and efficiently

Known TODOs:
- [ ] Implement a better structure than a sorted `Vec` (maybe)
- [ ] Tests
- [ ] Performance Test / verify with larger Ks 

When ran this with a K=1000 I got a panic when creating the output batches (I think I need to optimize the output creation for dictionary batches). 

```
thread 'tokio-runtime-worker' panicked at 'offset overflow', /Users/alamb/.cargo/registry/src/index.crates.io-6f17d22bba15001f/arrow-data-45.0.0/src/transform/utils.rs:42:56
stack backtrace:
   0: rust_begin_unwind
             at /rustc/8ede3aae28fe6e4d52b38157d7bfe0d3bceef225/library/std/src/panicking.rs:593:5
   1: core::panicking::panic_fmt
             at /rustc/8ede3aae28fe6e4d52b38157d7bfe0d3bceef225/library/core/src/panicking.rs:67:14
   2: core::panicking::panic_display
             at /rustc/8ede3aae28fe6e4d52b38157d7bfe0d3bceef225/library/core/src/panicking.rs:150:5
   3: core::panicking::panic_str
             at /rustc/8ede3aae28fe6e4d52b38157d7bfe0d3bceef225/library/core/src/panicking.rs:134:5
   4: core::option::expect_failed
             at /rustc/8ede3aae28fe6e4d52b38157d7bfe0d3bceef225/library/core/src/option.rs:1932:5
   5: core::option::Option<T>::expect
             at /rustc/8ede3aae28fe6e4d52b38157d7bfe0d3bceef225/library/core/src/option.rs:898:21
   6: arrow_data::transform::utils::extend_offsets::{{closure}}
             at /Users/alamb/.cargo/registry/src/index.crates.io-6f17d22bba15001f/arrow-data-45.0.0/src/transform/utils.rs:42:56
   7: core::iter::traits::iterator::Iterator::for_each::call::{{closure}}
             at /rustc/8ede3aae28fe6e4d52b38157d7bfe0d3bceef225/library/core/src/iter/traits/iterator.rs:853:29
   8: core::iter::traits::iterator::Iterator::fold
             at /rustc/8ede3aae28fe6e4d52b38157d7bfe0d3bceef225/library/core/src/iter/traits/iterator.rs:2481:21
   9: core::iter::traits::iterator::Iterator::for_each
             at /rustc/8ede3aae28fe6e4d52b38157d7bfe0d3bceef225/library/core/src/iter/traits/iterator.rs:856:9
  10: arrow_data::transform::utils::extend_offsets
             at /Users/alamb/.cargo/registry/src/index.crates.io-6f17d22bba15001f/arrow-data-45.0.0/src/transform/utils.rs:36:5
  11: arrow_data::transform::variable_size::build_extend::{{closure}}
             at /Users/alamb/.cargo/registry/src/index.crates.io-6f17d22bba15001f/arrow-data-45.0.0/src/transform/variable_size.rs:57:13
  12: <alloc::boxed::Box<F,A> as core::ops::function::Fn<Args>>::call
             at /rustc/8ede3aae28fe6e4d52b38157d7bfe0d3bceef225/library/alloc/src/boxed.rs:1999:9
  13: arrow_data::transform::MutableArrayData::extend
             at /Users/alamb/.cargo/registry/src/index.crates.io-6f17d22bba15001f/arrow-data-45.0.0/src/transform/mod.rs:631:9
  14: arrow_data::transform::MutableArrayData::with_capacities
             at /Users/alamb/.cargo/registry/src/index.crates.io-6f17d22bba15001f/arrow-data-45.0.0/src/transform/mod.rs:553:29
  15: arrow_data::transform::MutableArrayData::new
             at /Users/alamb/.cargo/registry/src/index.crates.io-6f17d22bba15001f/arrow-data-45.0.0/src/transform/mod.rs:353:9
  16: arrow_select::interleave::interleave_fallback
             at /Users/alamb/.cargo/registry/src/index.crates.io-6f17d22bba15001f/arrow-select-45.0.0/src/interleave.rs:198:26
  17: arrow_select::interleave::interleave
             at /Users/alamb/.cargo/registry/src/index.crates.io-6f17d22bba15001f/arrow-select-45.0.0/src/interleave.rs:90:14
  18: datafusion::physical_plan::topk::TopKHeap::emit::{{closure}}
```

Open Questions:
1. How / should we hook this into the planner? It could save multi-way sorting / merging 

## Performance testing

I tested with this dataset: [traces.zip](https://drive.google.com/file/d/1SlBGlTrYgcZTITDX_VRG22QZ7IyPX_JR/view?usp=drive_link) (240MB):

with a release build `cargo build --release`

I used two queries with `k=10` and `k=1000` and 

```
select * from traces order by time desc limit 10
# also ordering by high cardinality coumn:
select * from traces order by time desc, span_id limit 10
```


#### Performance

Using a high cardinality dictionary column:
```shell
select * from traces order by time desc, span_id limit 10
```


Using a high cardinality dictionary column:
```shell

```

| query | main(s) | this branch(s) | improvement |
|--------|--------|--------|--------|
| `select * from traces order by time desc limit 10`                         | 1.791 | 0.801 |  2.2 x faster|
| `select * from traces order by time desc, span_id limit 10` *     | 66.333. | 3.043 | 21 x faster |
| `select * from traces order by time desc limit 10000`                         |    |        |  3.393 (see above) |
| `select * from traces order by time desc, span_id limit 10000`                         |  3.291  |  DNF (see above)      |  XXX|


For the queries where this branch is 20x faster -- I didn't profile master, however, a noticed large amount of this time is spent using only a single core, which I believe is some sort of merge inside the `ExternalSort` operator.




#### Memory Usage
![Screenshot 2023-08-10 at 1 21 16 PM](https://github.com/apache/arrow-datafusion/assets/490673/be0fd2e5-0660-416b-ad6e-b486318ebe11)
![Screenshot 2023-08-10 at 1 23 25 PM](https://github.com/apache/arrow-datafusion/assets/490673/2c3035c4-a72a-4a1e-bb71-0bf9aab60d22)