@alamb 2023-08-13
I plan to try the following things with this PR, likely tomorrow:

- [ ] update storage to use `Vec<u8>` rather than `OwnedRow` (to avoid per row allocation)
- [ ] Test using  BTreeSet (to handle high row count)
- [ ] Implement a "compact" type operation for `RecordBatchStorage` to reduce memory needs