@alamb 2023-08-16
My next plan is to implement record batch compaction and then once I have that test out the performance of this branch with large values of K and adversarial input (reverse sorted input)