@alamb 2023-08-24
ok, I think I am now happy with the performance and feature set of this branch. I need to work out some interleave + dictionary nonsense with @tustvold  but I will begin trying to create a real PR shortly